criação do compoente Clock e reorganização das pastas
