+++ src/pages/login & register/login.jsx
@@ -2,9 +2,10 @@ import React from "react"
 import { Link } from "react-router-dom"
 
 // import RadioButton from "../../components/radioButton/radioButton"
-import ClickButon from "../../components/clickButton/clickButton"
-import InputTxt from "../../components/inputTxt/inputTxt"
+import ClickButon from "../../components/buttons&inputs/clickButton/clickButton"
+import InputTxt from "../../components/buttons&inputs/inputTxt/inputTxt"
 import LoggedUser from "../../services/loggedUser"
+import Header from "../../components/Time/clock"
 
 import "./login&register.css"
 import "../../globalStyle/colors.css"
@@ -12,9 +13,13 @@ import logoBurguerTiaNena from "../../images/logoBurguerTiaNena.png"
 
 function Login() {
     const {handleInputChange, loginSubmit, msgError} = LoggedUser();
+    const {seconds} = Header()
 
     return (
         <div className="container">
+            <div>
+            {seconds}
+                </div>
             <div className="logo-image">
                 <img className="logo-image" src={logoBurguerTiaNena} alt="logo"></img>
             </div>
